+++ packages/mobile/src/screens/notifications-screen/Notifications/index.ts
@@ -32,5 +32,5 @@ export * from './CommentThreadNotification'
 export * from './CommentMentionNotification'
 export * from './CommentReactionNotification'
 export * from './FanRemixContestEndedNotification'
-export * from './RemixContestEndingSoonNotification'
+export * from './FanRemixContestEndingSoonNotification'
 export * from './ArtistRemixContestEndedNotification'
